Fix bugs

+++ main.py
@@ -8,11 +8,11 @@ TEMPLATES = {
     "grpc": {"template_path": "templates/grpc", "extras_path": "extras/fastapi"},
     "ai-rest": {"template_path": "templates/ai-rest", "extras_path": "extras/fastapi"},
     "ai-grpc": {"template_path": "templates/ai-grpc", "extras_path": "extras/fastapi"},
-    "rest-grpc": {"template_path": "templates/rest_grpc", "extras_path": "extras/fastapi"},
-    "ai-rest-grpc": {"template_path": "templates/ai-rest_grpc", "extras_path": "extras/fastapi"},
+    "rest-grpc": {"template_path": "templates/rest-grpc", "extras_path": "extras/fastapi"},
+    "ai-rest-grpc": {"template_path": "templates/ai-rest-grpc", "extras_path": "extras/fastapi"},
     "data-science": {"template_path": "templates/data-science", "extras_path": "extras/data-science"},
-    "data-science-ml": {"template_path": "templates/data-science", "extras_path": "extras/data-science"},
-    "data-science-ml-ai": {"template_path": "templates/data-science", "extras_path": "extras/fastapi"},
+    "data-science-ml": {"template_path": "templates/data-science-ml", "extras_path": "extras/data-science"},
+    "data-science-ml-ai": {"template_path": "templates/data-science-ml-ai", "extras_path": "extras/fastapi"},
 }
 
 
@@ -34,6 +34,12 @@ def cli(template, output):
         click.echo(f"❌ Template '{template}' not found.")
         return
 
+    # Check if the destination directory already exists
+    if dst.exists():
+        click.echo(
+            f"❌ Directory '{dst}' already exists. Please choose a different output directory.")
+        return
+
     # Scaffold the project with the selected template
     click.echo(f"🚀 Scaffolding project with '{template}' template into {dst}")
     shutil.copytree(src, dst, dirs_exist_ok=True)
@@ -41,11 +47,15 @@ def cli(template, output):
     # Copy extras if they exist
     if extras.exists():
         click.echo(f"📦 Adding extras from {extras}...")
-        shutil.copytree(extras, dst / "extras", dirs_exist_ok=True)
+        extras_dst = dst / "extras"
+        # Ensure the extras folder exists
+        extras_dst.mkdir(parents=True, exist_ok=True)
+        shutil.copytree(extras, extras_dst, dirs_exist_ok=True)
 
     click.echo("\n✅ Done!")
     click.echo(
-        f"\n📦 To install dependencies:\n  pip install create_woragis_api[{template}]")
+        f"\n📦 To install dependencies:\n  pip install create_woragis_api[{template}]"
+    )
 
 
 if __name__ == "__main__":
